# swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Metadata": {
				"properties": {
					"lastViewed": {
						"type": "string"
					},
					"created": {
						"type": "string"
					},
					"timeToCook": {
						"type": "string"
					}
				},
				"required": [
					"lastViewed",
					"created",
					"timeToCook"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Steps": {
				"properties": {
					"step": {
						"type": "number",
						"format": "double"
					},
					"directions": {
						"type": "string"
					}
				},
				"required": [
					"step",
					"directions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Recipe": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"thumbnail": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"link": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"metadata": {
						"$ref": "#/components/schemas/Metadata"
					},
					"ingredients": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"steps": {
						"items": {
							"$ref": "#/components/schemas/Steps"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"thumbnail",
					"description",
					"link",
					"metadata",
					"ingredients",
					"steps"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "recipe-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/recipe": {
			"get": {
				"operationId": "GetAllRec",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {},
											"type": "array"
										},
										{
											"properties": {
												"message": {}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateRec",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Recipe"
							}
						}
					}
				}
			}
		},
		"/recipe/{recipeId}": {
			"get": {
				"operationId": "GetRec",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {},
											"type": "array"
										},
										{
											"properties": {
												"message": {}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "recipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteRec",
				"responses": {
					"202": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "recipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateRec",
				"responses": {
					"200": {
						"description": "Edited",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "recipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Recipe"
							}
						}
					}
				}
			}
		},
		"/recipe/search/{ingredient}": {
			"get": {
				"operationId": "SearchRec",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {},
											"type": "array"
										},
										{
											"properties": {
												"message": {}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "ingredient",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}